docs + better code organization
